refactor code to make it follow more like a std python lib stuct
